fix: Remove old social media login API
