--- tests/social-login/controller/callback-opener.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <script type="text/javascript">
-
-        function onLogin(msg){
-            console.log("*************", msg)
-        }
-
-        function popupWindow(url, title, win, w, h) {
-            const y = win.top.outerHeight / 2 + win.top.screenY - (h / 2);
-            const x = win.top.outerWidth / 2 + win.top.screenX - (w / 2);
-            return win.open(url, title, 'toolbar=no, location=no, directories=no, status=no, menubar=no, scrollbars=no, resizable=no, copyhistory=no, width=' + w + ', height=' + h + ', top=' + y + ', left=' + x);
-        }
-        
-        function openPopup() {
-            popupWindow("/view/popup", "Social Login", window, 200, 100)
-            return false;
-        }
-    </script>
-</head>
-
-<body>
-    <button onclick="openPopup()">Open</button>
-</body>
-
-</html>
