--- tests/social-login/controller/facebook-controller.ts
@@ -1,21 +1,0 @@
-import {
-    FacebookDialogProvider,
-    FacebookLoginStatus,
-    FacebookProvider,
-    oAuthCallback,
-    oAuthDialogEndPoint,
-} from "@plumier/social-login"
-import { bind, val } from "plumier"
-
-import { fb } from "../config"
-
-
-export class FacebookController {
-    @oAuthDialogEndPoint(new FacebookDialogProvider("/facebook/callback", fb.appId))
-    login() { }
-
-    @oAuthCallback(new FacebookProvider(fb.appId, fb.appSecret))
-    callback(@bind.loginStatus() profile: FacebookLoginStatus) {
-        return profile
-    }
-}
