--- tests/social-login/manual/view.ts
@@ -1,12 +1,0 @@
-import { join } from "path"
-import Plumier, { WebApiFacility } from "plumier"
-
-async function start(){
-    const callback = await new Plumier()
-            .set({ mode: "production" })
-            .set(new WebApiFacility({controller: join(__dirname, "../controller")}))
-            .initialize()
-     callback.listen(9000)
-}
-
-start()
