+++ tests/social-login/oauth2.spec.ts
